refactor: new file for report searializers and constants
KK-1026.

Move the ChildSerializer to a new serializers file, so the api.py does
not become too huge when more api views and more serializers are added.
Also move the serializer constants to a new conts.py file.
